# Improve performance of the index selector