--- a/crates/rsonpath-lib/src/automaton/state.rs
+++ b/crates/rsonpath-lib/src/automaton/state.rs
@@ -18,6 +18,11 @@ pub(crate) enum StateAttribute {
     /// Marks that the [`State`] contains some transition
     /// (labelled or fallback) to an [`Accepting`](`StateAttribute::Accepting`) state.
     TransitionsToAccepting = 0x08,
+    /// Marks that the [`State`] contains some transition labelled with an array index.
+    HasArrayIndexTransition = 0x10,
+    /// Marks that the [`State`] contains an array-index labelled transition
+    /// to an to an [`Accepting`](`StateAttribute::Accepting`) state.
+    HasArrayIndexTransitionToAccepting = 0x20,
 }
 
 pub(crate) struct StateAttributesBuilder {
@@ -49,6 +54,14 @@ impl StateAttributesBuilder {
 
     pub(crate) fn transitions_to_accepting(self) -> Self {
         self.set(StateAttribute::TransitionsToAccepting)
+    }
+
+    pub(crate) fn has_array_index_transition(self) -> Self {
+        self.set(StateAttribute::HasArrayIndexTransition)
+    }
+
+    pub(crate) fn has_array_index_transition_to_accepting(self) -> Self {
+        self.set(StateAttribute::HasArrayIndexTransitionToAccepting)
     }
 
     pub(crate) fn build(self) -> StateAttributes {
@@ -93,6 +106,12 @@ impl StateAttributes {
     /// A state is _unitary_ if it contains exactly one labelled transition
     /// and its fallback transition is [`Rejecting`](`StateAttributes::is_rejecting`).
     pub const UNITARY: Self = Self(StateAttribute::Unitary as u8);
+    /// Marks that the [`State`] contains some transition labelled with an array index.
+    pub const HAS_ARRAY_INDEX_TRANSITION: Self = Self(StateAttribute::HasArrayIndexTransition as u8);
+    /// Marks that the [`State`] contains an array-index labelled transition
+    /// to an to an [`Accepting`](`StateAttributes::is_accepting`) state.
+    pub const HAS_ARRAY_INDEX_TRANSITION_TO_ACCEPTING: Self =
+        Self(StateAttribute::HasArrayIndexTransitionToAccepting as u8);
 
     /// Check if the the state is accepting.
     #[inline(always)]
@@ -124,6 +143,21 @@ impl StateAttributes {
     #[must_use]
     pub fn is_unitary(&self) -> bool {
         self.is_set(StateAttribute::Unitary)
+    }
+
+    /// Marks that the [`State`] contains some transition labelled with an array index.
+    #[inline(always)]
+    #[must_use]
+    pub fn has_array_index_transition(&self) -> bool {
+        self.is_set(StateAttribute::HasArrayIndexTransition)
+    }
+
+    /// Marks that the [`State`] contains an array-index labelled transition
+    /// to an to an [`Accepting`](`StateAttributes::is_accepting`) state.
+    #[inline(always)]
+    #[must_use]
+    pub fn has_array_index_transition_to_accepting(&self) -> bool {
+        self.is_set(StateAttribute::HasArrayIndexTransitionToAccepting)
     }
 
     #[inline(always)]