… crates/rsonpath-lib/src/automaton/minimizer.rs