## Short description

Improved the access patterns for array index selectors.

First, member and array transitions are now in separate smallvec arrays, so we can always look only at the relevant array transitions.

Second, properties like "has array transition" or "can accept via array transition" are now cached during `minimize` as `StateAttributes`.

Overall on chirop@lille this gives from 7% to 12% thpt increase on most of the `main` benchmark.
See full bench report at the bottom.

## Issue

Resolves: #138 
Rel: #151 

## Checklist

All of these should be ticked off before you submit the PR.

- [x] I ran `just verify` locally and it succeeded.
- [x] Issue was given <span style="color: #FF4400">go ahead</span> and is linked above **OR** I have included justification for a minor change.
- [x] Unit tests for my changes are included **OR** no functionality was changed.

## Benchmark report

There is a weird regression on `./data/twitter/twitter.json_user_mentions_indices/rsonpath_mmap_count` which should be investigated as part of #151. However, the change looks reasonably beneficial on all other benches.

```
./data/nativejson/canada.json_canada::second_coord_component/rsonpath
                        time:   [5.0144 ms 5.0161 ms 5.0178 ms]
                        thrpt:  [448.61 MB/s 448.77 MB/s 448.92 MB/s]
                 change:
                        time:   [-5.7562% -5.7026% -5.6470%] (p = 0.00 < 0.05)
                        thrpt:  [+5.9850% +6.0474% +6.1078%]
                        Performance has improved.
./data/nativejson/canada.json_canada::second_coord_component/rsonpath_count
                        time:   [2.2977 ms 2.2989 ms 2.3004 ms]
                        thrpt:  [978.57 MB/s 979.18 MB/s 979.69 MB/s]
                 change:
                        time:   [-6.9102% -6.7433% -6.4948%] (p = 0.00 < 0.05)
                        thrpt:  [+6.9460% +7.2308% +7.4232%]
                        Performance has improved.
Found 1 outliers among 10 measurements (10.00%)
  1 (10.00%) high severe
./data/nativejson/canada.json_canada::second_coord_component/rsonpath_mmap
                        time:   [5.1043 ms 5.1071 ms 5.1144 ms]
                        thrpt:  [440.14 MB/s 440.77 MB/s 441.01 MB/s]
                 change:
                        time:   [-7.7144% -7.6213% -7.5149%] (p = 0.00 < 0.05)
                        thrpt:  [+8.1256% +8.2500% +8.3593%]
                        Performance has improved.
Found 1 outliers among 10 measurements (10.00%)
  1 (10.00%) high mild
./data/nativejson/canada.json_canada::second_coord_component/rsonpath_mmap_count
                        time:   [2.3352 ms 2.3417 ms 2.3467 ms]
                        thrpt:  [959.22 MB/s 961.29 MB/s 963.96 MB/s]
                 change:
                        time:   [-11.937% -11.472% -11.059%] (p = 0.00 < 0.05)
                        thrpt:  [+12.434% +12.958% +13.555%]
                        Performance has improved.

./data/nativejson/canada.json_canada::coord_476_1446_1/rsonpath
                        time:   [1.5509 ms 1.5513 ms 1.5518 ms]
                        thrpt:  [1.4506 GB/s 1.4510 GB/s 1.4515 GB/s]
                 change:
                        time:   [-11.986% -11.870% -11.763%] (p = 0.00 < 0.05)
                        thrpt:  [+13.332% +13.469% +13.618%]
                        Performance has improved.
Found 1 outliers among 10 measurements (10.00%)
  1 (10.00%) high mild
./data/nativejson/canada.json_canada::coord_476_1446_1/rsonpath_count
                        time:   [1.2077 ms 1.2155 ms 1.2213 ms]
                        thrpt:  [1.8431 GB/s 1.8519 GB/s 1.8639 GB/s]
                 change:
                        time:   [-12.584% -12.304% -11.970%] (p = 0.00 < 0.05)
                        thrpt:  [+13.598% +14.031% +14.396%]
                        Performance has improved.
Found 1 outliers among 10 measurements (10.00%)
  1 (10.00%) high mild
./data/nativejson/canada.json_canada::coord_476_1446_1/rsonpath_mmap
                        time:   [1.6201 ms 1.6213 ms 1.6226 ms]
                        thrpt:  [1.3874 GB/s 1.3884 GB/s 1.3894 GB/s]
                 change:
                        time:   [-11.730% -11.653% -11.571%] (p = 0.00 < 0.05)
                        thrpt:  [+13.085% +13.190% +13.289%]
                        Performance has improved.
./data/nativejson/canada.json_canada::coord_476_1446_1/rsonpath_mmap_count
                        time:   [1.1908 ms 1.1921 ms 1.1931 ms]
                        thrpt:  [1.8868 GB/s 1.8883 GB/s 1.8903 GB/s]
                 change:
                        time:   [-15.611% -15.479% -15.355%] (p = 0.00 < 0.05)
                        thrpt:  [+18.140% +18.314% +18.498%]
                        Performance has improved.
Found 2 outliers among 10 measurements (20.00%)
  1 (10.00%) low mild
  1 (10.00%) high mild

./data/nativejson/citm.json_citm::seatCategoryId/rsonpath
                        time:   [375.17 µs 375.32 µs 375.43 µs]
                        thrpt:  [4.6007 GB/s 4.6019 GB/s 4.6037 GB/s]
                 change:
                        time:   [-3.8983% -3.1712% -2.4736%] (p = 0.00 < 0.05)
                        thrpt:  [+2.5363% +3.2751% +4.0564%]
                        Performance has improved.
./data/nativejson/citm.json_citm::seatCategoryId/rsonpath_count
                        time:   [233.44 µs 235.26 µs 237.55 µs]
                        thrpt:  [7.2708 GB/s 7.3418 GB/s 7.3989 GB/s]
                 change:
                        time:   [-16.799% -15.243% -13.734%] (p = 0.00 < 0.05)
                        thrpt:  [+15.920% +17.984% +20.191%]
                        Performance has improved.
Found 2 outliers among 10 measurements (20.00%)
  2 (20.00%) high mild
./data/nativejson/citm.json_citm::seatCategoryId/rsonpath_mmap
                        time:   [355.23 µs 356.24 µs 357.78 µs]
                        thrpt:  [4.8275 GB/s 4.8485 GB/s 4.8622 GB/s]
                 change:
                        time:   [+0.7732% +1.3615% +1.9851%] (p = 0.00 < 0.05)
                        thrpt:  [-1.9464% -1.3432% -0.7673%]
                        Change within noise threshold.
Found 1 outliers among 10 measurements (10.00%)
  1 (10.00%) high mild
./data/nativejson/citm.json_citm::seatCategoryId/rsonpath_mmap_count
                        time:   [218.55 µs 218.87 µs 219.34 µs]
                        thrpt:  [7.8747 GB/s 7.8916 GB/s 7.9030 GB/s]
                 change:
                        time:   [+1.5263% +1.7435% +1.9861%] (p = 0.00 < 0.05)
                        thrpt:  [-1.9475% -1.7137% -1.5033%]
                        Performance has regressed.

./data/ast/ast.json_ast::nested_inner/rsonpath
                        time:   [28.864 ms 28.876 ms 28.888 ms]
                        thrpt:  [885.95 MB/s 886.32 MB/s 886.68 MB/s]
                 change:
                        time:   [-2.9534% -2.8978% -2.8412%] (p = 0.00 < 0.05)
                        thrpt:  [+2.9242% +2.9843% +3.0432%]
                        Performance has improved.
Found 3 outliers among 10 measurements (30.00%)
  1 (10.00%) low mild
  2 (20.00%) high mild
./data/ast/ast.json_ast::nested_inner/rsonpath_count
                        time:   [20.616 ms 20.626 ms 20.637 ms]
                        thrpt:  [1.2402 GB/s 1.2408 GB/s 1.2414 GB/s]
                 change:
                        time:   [-3.9317% -3.7078% -3.5175%] (p = 0.00 < 0.05)
                        thrpt:  [+3.6458% +3.8505% +4.0926%]
                        Performance has improved.
Found 1 outliers among 10 measurements (10.00%)
  1 (10.00%) high mild
./data/ast/ast.json_ast::nested_inner/rsonpath_mmap
                        time:   [29.178 ms 29.187 ms 29.196 ms]
                        thrpt:  [876.61 MB/s 876.87 MB/s 877.13 MB/s]
                 change:
                        time:   [-4.0546% -4.0203% -3.9910%] (p = 0.00 < 0.05)
                        thrpt:  [+4.1569% +4.1887% +4.2260%]
                        Performance has improved.
./data/ast/ast.json_ast::nested_inner/rsonpath_mmap_count
                        time:   [20.391 ms 20.416 ms 20.435 ms]
                        thrpt:  [1.2524 GB/s 1.2536 GB/s 1.2551 GB/s]
                 change:
                        time:   [-8.6349% -8.1167% -7.7519%] (p = 0.00 < 0.05)
                        thrpt:  [+8.4034% +8.8337% +9.4510%]
                        Performance has improved.
Found 2 outliers among 10 measurements (20.00%)
  1 (10.00%) low mild
  1 (10.00%) high mild

./data/ast/ast.json_ast::deepest/rsonpath
                        time:   [17.790 ms 17.795 ms 17.806 ms]
                        thrpt:  [1.4373 GB/s 1.4382 GB/s 1.4387 GB/s]
                 change:
                        time:   [+0.0534% +0.1436% +0.2409%] (p = 0.01 < 0.05)
                        thrpt:  [-0.2403% -0.1434% -0.0533%]
                        Change within noise threshold.
./data/ast/ast.json_ast::deepest/rsonpath_count
                        time:   [13.634 ms 13.652 ms 13.676 ms]
                        thrpt:  [1.8714 GB/s 1.8747 GB/s 1.8771 GB/s]
                 change:
                        time:   [+1.0579% +1.2534% +1.4533%] (p = 0.00 < 0.05)
                        thrpt:  [-1.4325% -1.2379% -1.0468%]
                        Performance has regressed.
./data/ast/ast.json_ast::deepest/rsonpath_mmap
                        time:   [17.880 ms 17.891 ms 17.902 ms]
                        thrpt:  [1.4296 GB/s 1.4305 GB/s 1.4314 GB/s]
                 change:
                        time:   [-1.6591% -1.6058% -1.5529%] (p = 0.00 < 0.05)
                        thrpt:  [+1.5774% +1.6320% +1.6870%]
                        Performance has improved.
./data/ast/ast.json_ast::deepest/rsonpath_mmap_count
                        time:   [13.273 ms 13.323 ms 13.411 ms]
                        thrpt:  [1.9083 GB/s 1.9210 GB/s 1.9282 GB/s]
                 change:
                        time:   [-2.1778% -1.7634% -1.3030%] (p = 0.00 < 0.05)
                        thrpt:  [+1.3202% +1.7951% +2.2263%]
                        Performance has improved.
Found 1 outliers among 10 measurements (10.00%)
  1 (10.00%) high mild

./data/pison/bestbuy_short_record.json_bestbuy::products_category/rsonpath
                        time:   [34.178 ms 34.222 ms 34.271 ms]
                        thrpt:  [3.0122 GB/s 3.0166 GB/s 3.0204 GB/s]
                 change:
                        time:   [+1.4234% +1.5005% +1.5818%] (p = 0.00 < 0.05)
                        thrpt:  [-1.5572% -1.4784% -1.4035%]
                        Performance has regressed.
Found 1 outliers among 10 measurements (10.00%)
  1 (10.00%) high severe
./data/pison/bestbuy_short_record.json_bestbuy::products_category/rsonpath_count
                        time:   [26.736 ms 27.117 ms 28.026 ms]
                        thrpt:  [3.6834 GB/s 3.8069 GB/s 3.8612 GB/s]
                 change:
                        time:   [-4.8178% -1.5078% +1.1216%] (p = 0.36 > 0.05)
                        thrpt:  [-1.1091% +1.5309% +5.0616%]
                        No change in performance detected.
./data/pison/bestbuy_short_record.json_bestbuy::products_category/rsonpath_mmap
                        time:   [34.782 ms 34.791 ms 34.810 ms]
                        thrpt:  [2.9656 GB/s 2.9672 GB/s 2.9679 GB/s]
                 change:
                        time:   [+1.3360% +1.4133% +1.5023%] (p = 0.00 < 0.05)
                        thrpt:  [-1.4801% -1.3936% -1.3183%]
                        Performance has regressed.
./data/pison/bestbuy_short_record.json_bestbuy::products_category/rsonpath_mmap_count
                        time:   [28.898 ms 29.015 ms 29.126 ms]
                        thrpt:  [3.5443 GB/s 3.5579 GB/s 3.5722 GB/s]
                 change:
                        time:   [+1.4799% +1.8876% +2.3335%] (p = 0.00 < 0.05)
                        thrpt:  [-2.2803% -1.8526% -1.4583%]
                        Performance has regressed.

./data/pison/bestbuy_short_record.json_bestbuy::products_video_only/rsonpath_direct_count
                        time:   [109.04 ms 109.07 ms 109.09 ms]
                        thrpt:  [946.33 MB/s 946.50 MB/s 946.71 MB/s]
                 change:
                        time:   [-8.5134% -8.4821% -8.4559%] (p = 0.00 < 0.05)
                        thrpt:  [+9.2370% +9.2683% +9.3056%]
                        Performance has improved.
./data/pison/bestbuy_short_record.json_bestbuy::products_video_only/rsonpath_descendant_count
                        time:   [9.5149 ms 9.5809 ms 9.6211 ms]
                        thrpt:  [10.730 GB/s 10.775 GB/s 10.850 GB/s]
                 change:
                        time:   [-2.3945% -1.9227% -1.4300%] (p = 0.00 < 0.05)
                        thrpt:  [+1.4508% +1.9604% +2.4533%]
                        Performance has improved.
./data/pison/bestbuy_short_record.json_bestbuy::products_video_only/rsonpath_direct_nodes
                        time:   [135.55 ms 135.62 ms 135.67 ms]
                        thrpt:  [760.92 MB/s 761.17 MB/s 761.57 MB/s]
                 change:
                        time:   [-4.2566% -4.1786% -4.1099%] (p = 0.00 < 0.05)
                        thrpt:  [+4.2860% +4.3608% +4.4458%]
                        Performance has improved.
./data/pison/bestbuy_short_record.json_bestbuy::products_video_only/rsonpath_descendant_nodes
                        time:   [12.185 ms 12.193 ms 12.206 ms]
                        thrpt:  [8.4577 GB/s 8.4666 GB/s 8.4723 GB/s]
                 change:
                        time:   [-6.1811% -5.7272% -5.2631%] (p = 0.00 < 0.05)
                        thrpt:  [+5.5555% +6.0751% +6.5884%]
                        Performance has improved.

./data/pison/bestbuy_short_record.json_bestbuy::all_nodes/rsonpath
                        time:   [691.68 ms 699.48 ms 709.60 ms]
                        thrpt:  [145.48 MB/s 147.58 MB/s 149.25 MB/s]
                 change:
                        time:   [-2.5324% -0.9290% +0.9107%] (p = 0.33 > 0.05)
                        thrpt:  [-0.9024% +0.9377% +2.5982%]
                        No change in performance detected.
Found 1 outliers among 10 measurements (10.00%)
  1 (10.00%) high mild
./data/pison/bestbuy_short_record.json_bestbuy::all_nodes/rsonpath_count
                        time:   [87.599 ms 87.647 ms 87.709 ms]
                        thrpt:  [1.1770 GB/s 1.1778 GB/s 1.1785 GB/s]
                 change:
                        time:   [-2.8238% -2.7557% -2.6838%] (p = 0.00 < 0.05)
                        thrpt:  [+2.7578% +2.8338% +2.9059%]
                        Performance has improved.
./data/pison/bestbuy_short_record.json_bestbuy::all_nodes/rsonpath_mmap
                        time:   [705.00 ms 708.70 ms 712.45 ms]
                        thrpt:  [144.90 MB/s 145.66 MB/s 146.43 MB/s]
                 change:
                        time:   [-0.6276% +0.3405% +1.2947%] (p = 0.52 > 0.05)
                        thrpt:  [-1.2782% -0.3394% +0.6316%]
                        No change in performance detected.
./data/pison/bestbuy_short_record.json_bestbuy::all_nodes/rsonpath_mmap_count
                        time:   [94.240 ms 94.306 ms 94.371 ms]
                        thrpt:  [1.0939 GB/s 1.0946 GB/s 1.0954 GB/s]
                 change:
                        time:   [-3.4564% -3.3477% -3.2443%] (p = 0.00 < 0.05)
                        thrpt:  [+3.3531% +3.4637% +3.5801%]
                        Performance has improved.

./data/pison/google_map_short_record.json_google_map::routes/rsonpath
                        time:   [60.286 ms 60.296 ms 60.312 ms]
                        thrpt:  [1.7724 GB/s 1.7729 GB/s 1.7732 GB/s]
                 change:
                        time:   [-2.4644% -2.4199% -2.3758%] (p = 0.00 < 0.05)
                        thrpt:  [+2.4336% +2.4799% +2.5267%]
                        Performance has improved.
./data/pison/google_map_short_record.json_google_map::routes/rsonpath_count
                        time:   [48.933 ms 48.960 ms 48.981 ms]
                        thrpt:  [2.1824 GB/s 2.1833 GB/s 2.1846 GB/s]
                 change:
                        time:   [-3.7676% -3.1628% -2.6783%] (p = 0.00 < 0.05)
                        thrpt:  [+2.7520% +3.2661% +3.9151%]
                        Performance has improved.
Found 2 outliers among 10 measurements (20.00%)
  2 (20.00%) high mild
./data/pison/google_map_short_record.json_google_map::routes/rsonpath_mmap
                        time:   [60.411 ms 60.429 ms 60.441 ms]
                        thrpt:  [1.7686 GB/s 1.7690 GB/s 1.7695 GB/s]
                 change:
                        time:   [-4.9174% -4.8854% -4.8553%] (p = 0.00 < 0.05)
                        thrpt:  [+5.1031% +5.1364% +5.1717%]
                        Performance has improved.
./data/pison/google_map_short_record.json_google_map::routes/rsonpath_mmap_count
                        time:   [48.279 ms 48.293 ms 48.321 ms]
                        thrpt:  [2.2122 GB/s 2.2135 GB/s 2.2141 GB/s]
                 change:
                        time:   [-5.1179% -5.0657% -5.0165%] (p = 0.00 < 0.05)
                        thrpt:  [+5.2814% +5.3360% +5.3940%]
                        Performance has improved.

./data/pison/google_map_short_record.json_google_map::travel_modes/rsonpath_direct_count
                        time:   [16.981 ms 17.003 ms 17.024 ms]
                        thrpt:  [6.2790 GB/s 6.2869 GB/s 6.2953 GB/s]
                 change:
                        time:   [-1.6973% -1.5067% -1.3452%] (p = 0.00 < 0.05)
                        thrpt:  [+1.3636% +1.5297% +1.7266%]
                        Performance has improved.
./data/pison/google_map_short_record.json_google_map::travel_modes/rsonpath_descendant_count
                        time:   [9.9101 ms 9.9382 ms 9.9941 ms]
                        thrpt:  [10.696 GB/s 10.756 GB/s 10.787 GB/s]
                 change:
                        time:   [-10.888% -10.255% -9.7726%] (p = 0.00 < 0.05)
                        thrpt:  [+10.831% +11.427% +12.218%]
                        Performance has improved.
./data/pison/google_map_short_record.json_google_map::travel_modes/rsonpath_direct_nodes
                        time:   [18.157 ms 18.177 ms 18.206 ms]
                        thrpt:  [5.8716 GB/s 5.8810 GB/s 5.8874 GB/s]
                 change:
                        time:   [-1.4163% -1.3285% -1.2425%] (p = 0.00 < 0.05)
                        thrpt:  [+1.2581% +1.3464% +1.4367%]
                        Performance has improved.
./data/pison/google_map_short_record.json_google_map::travel_modes/rsonpath_descendant_nodes
                        time:   [12.299 ms 12.301 ms 12.303 ms]
                        thrpt:  [8.6883 GB/s 8.6903 GB/s 8.6917 GB/s]
                 change:
                        time:   [-1.6944% -1.3742% -1.1106%] (p = 0.00 < 0.05)
                        thrpt:  [+1.1231% +1.3934% +1.7236%]
                        Performance has improved.

./data/ast/ast.json_inner_array/rsonpath
                        time:   [57.144 ms 57.256 ms 57.330 ms]
                        thrpt:  [446.42 MB/s 447.00 MB/s 447.87 MB/s]
                 change:
                        time:   [-5.4647% -5.2995% -5.1376%] (p = 0.00 < 0.05)
                        thrpt:  [+5.4159% +5.5960% +5.7806%]
                        Performance has improved.
./data/ast/ast.json_inner_array/rsonpath_count
                        time:   [17.636 ms 17.646 ms 17.655 ms]
…